pyproject.toml
[project]
name = "symbiosis"
version = "0.1.0"
description = "Multi-layered LLM agent framework"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "openai>=1.0",
    "anthropic>=0.40",
    "croniter>=2.0",
]

[project.optional-dependencies]
neural = [
    "torch>=2.0",
]
rendering = [
    "matplotlib>=3.7",
    "Pillow>=10.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
symbiosis = "library.__main__:main"
matrix = "library.matrix_cli:main"

[tool.hatch.build.targets.wheel]
packages = ["library"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
